Add coverage per codon column support when parsing input TSV
The program calculate_codon_frequencies.py now output an additional
column with coverage counts for each codon site as parsed from the reference
sequence. When '---' is in the reference then nucleotides parsed from the sample
in the "codon" are considered INSertion. Incomplete codons in the sample are now
also recognized and counted as, e.g. 'GG-' and output in the TSV file (in such
cases 'X' is output as the amino acid residue.