Add support for detection of INSertion events relative to reference
So far INSertions could not be reported because the reference would
need to be padded with dashes depending where the INSertions actually
appear in sample reads. One can however manually extend the reference
sequence with dashes as needed. Further, changes to Biopython
Bio.Seq.translate() needed to be developed (unsubmitted yet). Now, the
codon '---' translates to '-'.

There is still one caveat: The inserted codons and aminoacids obtain
same number as the preceding codon/amino acids residue, like implemented
in PDB and elsewhere. The problem is that when drawing figures the
INSerted amino acid or codon shows in the very same column as the one it
is preceding by. Hence, they are mixed in the very same column of the
X-axis.

Drop support for parsing read count from the below entry as the number
has a different meaning:

    >BA2.86_(222)
    atgtaa