Refactored code to require proper alignment to support INSertions
So far the code kept picking more characters until it collected three for a full
codon to be translatable. We now require the input to be fully aligned. That means
if 3nt DELetion occurred across two codon both will be translated into X although
reading-frame was preserved. Users must edit the alignment and align the nucleotides
to occupy either of the two codons so that either of them will appear as changed
and the other one as DELeted. Look for adjacent XX in translated sequences into
protein or for successive X in the col4.

To facilitate display of INSerted codons lets add an extra column to the TSV output
where all codon of the input aligned are numbered consequently (col1).
In col2 we keep the number of the original immediately preceding codon like is generally
recognized, so that INSerted codon retains a position of the one preceeding it.
So in col2 one can see the original numbering of the codons and amino acid residues
as in the (unpadded) reference.

In col8 we keep absolute coverage per each codon.

$ grep '^48[23]' tests/outputs/test.unchanged_codons.tsv
482	481	N	N	1.000000	AAT	AAT	10
483	482	G	G	0.200000	GGT	GGT	10
$ cat tests/outputs/test.tsv
2	1	INS	P	0.100000	---	CCC	10
483	482	G	X	0.800000	GGT	GG	10
484	483	V	X	0.800000	GTT	T	10
491	490	F	K	0.100000	TTT	AAA	10
1273	1272	Y	DEL	0.100000	TAC	---	9
$

Drop requirement on bcbi-gff parses which was not really necessary.

Add some minimalistic testcases, they are already documented in the README.md.