Release 0.2
This release can process FASTA entries with the counts of unified sequences
and aligned by NCBI blastn for maximum quality and width of the alignment.
To support calling DELetion and INSertion events it requires the aligned
sequences to be of same lengths as the reference. That means that if in the
sample are any INSertions then the reference must be inflated by padding
dashes accordingly to have same length.
In our experiments we focus on nucleotide mismatches so this was so far not
an issue for us so far, especially because any INS or DEL events (NGS
sequencing errors) would typically break the ribosomal reading frame. But the
encoded proteins were expressed on the surface of our yeast cells so we know
the INS or DEL breaking the reading frame are just errors. In our setup.
However, we observed some DEL events spanning multiples of 3 nucleotides so
we made the software more robust.

The TSV files now have two additional columns to keep track of the number of
reads containing a mutated codon and also, to the total number of reads
spanning the codon site (total coverage).

Sometimes it is handy to disable 2nd Y-axis.

Better document cmdline options and their defaults.