Release 0.2
This release can process FASTA with the counts of unified sequences and aligned by
NCBI blastn for maximum quality and width of the alignment. To support calling
DELetion and INSertion events it requires the aligned sequences to be of same lengths
as the reference. That means that if in the sample are any INSertions then the reference
must be inflated by padding dashes accordingly. In our experiments we focus on
nucleotide mismatches so this was so far not an issue for us, especially if any
INS or DEL event deemed from NGS sequence reads would typically break the ribosomal
reading frame. Those are just sequencing errors in our setup. However, we observed
some DEL events spanning multiples of 3 nucleotides so we made the software more robust.

The TSV file now have two additional columns to keep track of the number of reads
containing a mutated codon and also, to the total number of reads spanning the codon
site (total coverage).